Remove SMTP-based tests.
These tests attempted to verify that email sent to a given email
address would be delivered. Given the reality of today's Internet,
this is fundamentally impossible. The exact thing we tried to do,
verifying addresses, is actively being fought by mail server
operators, for reasons of spam-reduction and privacy. There were many
known (and commonly occuring in the wild) cases where our tests would
give the wrong result, both as false positives and false negatives.
Rather than waste time and resources on tests that don't tell us
anything, we will in the future limit ourselves to checking that
addresses are syntactically valid and that the domain they point at
has a designated mail server.